Implement StatusController and StatusClient class
StatusController and StatusClient are in charge of controlling access to the
statusbar and progressbar. In particular, external classes may only access the
statusbar and progressbar through the StatusClient which is issued by
StatusController.

StatusController in turn makes sure that only one client may exist per
controller at any point in time.