Binpacker: Use ratio of aspect ratios as metric
A normalized (0.0-1.0) ratio of aspect ratios is a fairer comparison metric for
comparing ratios. For example, comparing a 4:5 and 5:4 box with a 1:1 ratio
using the difference in aspect ratios will result in:
 - 4:5 -- abs(0.8 - 1.0) = 0.2
 - 5:4 -- abs(1.25 - 1.0) = 0.25

The results should be same, as both should be considered equallty close to a
square aspect ratio. Using the new ratio of ratios method, however:
 - 4:5 -- 0.8 / 1.0 = 0.8
 - 5:4 -- 1.0 / 1.25 = 0.8

The result is a fairer comparison of real vs target aspect ratios, resulting in
a slightly better heuristic.