Rectangles: Unconditionally call recalculate_size_impl()
When recalculate_size() is triggered at a certain level in the tree, we need to
recalculate it before going up the tree to the root.

This is because aspect_ratio() starts flapping around due to width()/height()
changes after a layout change, causing some rather non-deterministic behaviour
when recalculating the size.